weather component added
