+++ frontend/src/Components/condition.css
@@ -2,11 +2,12 @@
     background: rgba(255, 255, 255, 0.3);
     backdrop-filter: blur(10px);
     border-radius: 15px;
-    padding: 16px;
+    padding: 6px;
     display: flex;
     align-items: center;
     justify-content: space-between;
-    margin-bottom: 16px;
+    display: flex;
+    flex-direction: column;
     box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
     font-family: 'Outfit', BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
         'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
@@ -21,13 +22,13 @@
 }
 
 .condition-attribute-name {
-    color: white;
+    color: rgb(0, 0, 0);
     
     font-weight: 600;
     
 }
 
 .condition-attribute-value {
-    color: white;
+    color: rgb(0, 0, 0);
     font-weight: 600;
 }
