+++ frontend/src/Components/weather.css
@@ -1,92 +1,81 @@
-.top{
+.weatherComponent {
+     border-radius: 15px;
+    padding: 20px;
+     width: 50%;
+    margin: 20px ;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
     display: flex;
-    flex-direction: row;
-    margin-bottom: 10px;
-    
-    
-}
-
-.secondColumn{
-    display: flex;
-    flex: 1;
     flex-direction: column;
-    
+     
+
+
 }
 
-.location{
+ 
+
+.row1{
     display: flex;
+    flex-direction: column;
     flex: 2;
-    font-size: 3rem;
-    font-weight: 600;
-   
-
-}
-
-.temp{
-    display: flex;
-    justify-content: end;
-    
-  
-   
-
-}
-.tempText{
-    font-size: 1.4rem;
-    font-weight:400;
-   margin-bottom: 0;
-
-}
-.desText{
     font-size: 1rem;
-    font-weight:400;
-   margin-top: 0;
-
-   
-
+    color: #333;
 }
 
+.locationText {
+    font-size: 1.4rem;
+    font-weight: 600;
+    margin: 0;
+    color: #333;
+}
 
-.locationText{
+.subText {
+    font-size: 0.9rem;
+    color: #777;
+    margin: 2px 0;
+}
+
+.row2 {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    flex: 1;
+}
+
+.solar{
+    font-size: 3rem;
+    font-weight: 700;
+    color: #333;
+    margin-right: 10px;
+    flex: 5;
+}
+.description{
+    flex:1;
+}
+
+.solarText{
     font-size: 2.5rem;
     font-weight: 600;
-    color: #000000;
-    margin: 0;
+    color: #333;
 }
 
-.description{
-    
-    padding-top: 0;
-    height: min-content;
+.icon {
+    width: 50px;
+    height: 50px;
+}
 
+.row3 {
     display: flex;
-    justify-content: end;
-    
-
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: space-around;
+    margin-top: 20px;
 }
 
-
-
-.search{
-    text-align: center;
-    padding: 0;
-     display: flex;
-     margin: 2px;
-     margin-bottom: 5px;
- }
-
-input{
-    padding: 0.7rem;
+.attribute {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 10px;
     font-size: 1rem;
-    border-radius: 5px;
-    border: 1px solid #000000;
-    
-    width: 100%;
-    border-radius: 25px;
-    border: 2px solid rgba(255,255,255,0.3);
-    background: rgba(255,255,255,0.3);
-
-}
-
-::placeholder{
-     color: #f8f8f8
+    color: #333;
 }
