+++ frontend/src/Pages/Dashboard.jsx
@@ -2,14 +2,13 @@ import React, { useState } from "react";
 import backgroundImage from '../Components/Assets/Background.png';
 import { Link } from "react-router-dom";
 import Header from "../Components/header";
-import SearchBar from "../Components/searchBar";
-import Weather from "../Components/weather";
+ import Weather from "../Components/weather";
 
 function DashboardPage() {
   return (
     <div style={styles.container}>
       <Header />
-      <SearchBar />
+     
       <Weather />
       
     </div>
