+++ frontend/src/weather.json
@@ -1,0 +1,37 @@
+{
+  "coord": { "lon": 79.8478, "lat": 6.9319 },
+  "weather": [
+    {
+      "id": 804,
+      "main": "Clouds",
+      "description": "overcast clouds",
+      "icon": "04d"
+    }
+  ],
+  "base": "stations",
+  "main": {
+    "temp": 304.12,
+    "feels_like": 311.12,
+    "temp_min": 304.12,
+    "temp_max": 304.34,
+    "pressure": 1012,
+    "humidity": 73,
+    "sea_level": 1012,
+    "grnd_level": 1011
+  },
+  "visibility": 10000,
+  "wind": { "speed": 4.31, "deg": 229, "gust": 5.34 },
+  "clouds": { "all": 100 },
+  "dt": 1724478948,
+  "sys": {
+    "type": 1,
+    "id": 9098,
+    "country": "LK",
+    "sunrise": 1724459659,
+    "sunset": 1724503920
+  },
+  "timezone": 19800,
+  "id": 1248991,
+  "name": "Colombo",
+  "cod": 200
+}
